토큰 클래스 추가 및 머지
토큰 사용을 편리하게 만들기 위해 토큰 클래스 추가
머지하다가 문제 생긴 부분 수정
